+++ scripts/macros.js
@@ -24,6 +24,7 @@ export {reduce, reduceReduced} from './macros/2014/monsterFeatures/generic/reduc
 export {reduceMaxHP} from './macros/2014/monsterFeatures/generic/reduceMaxHP.js';
 export {rerollSaveOnDamage, rerollSaveOnDamageActor} from './macros/2014/monsterFeatures/generic/rerollSaveOnDamage.js';
 export {spellTurning} from './macros/2014/monsterFeatures/generic/spellTurning.js';
+export {suffocate} from './macros/2014/monsterFeatures/generic/suffocate.js';
 export {sunlightSensitivity} from './macros/2014/monsterFeatures/generic/sunlightSensitivity.js';
 export {surpriseAttack} from './macros/2014/monsterFeatures/generic/surpriseAttack.js';
 export {swarmDamage} from './macros/2014/monsterFeatures/generic/swarmDamage.js';
@@ -32,6 +33,8 @@ export {undeadFortitude} from './macros/2014/monsterFeatures/generic/undeadForti
 //Mechanics
 export {cleaveMastery, grazeMastery, pushMastery, sapMastery, slowMastery, toppleMastery, vexMastery} from './macros/2024/mechanics/masteries.js';
 export {invisible} from './macros/2024/mechanics/invisible.js';
+//Actions
+export {suffocation, _suffocationApply, suffocationOutOfAir} from './macros/2024/actions/suffocation.js';
 //Generic
 export {proneOnFail} from './macros/2024/generic/proneOnFail.js';
 //Spells